ver. 1.5
ver. 1.5
- add sheet and operations pages
- add sum to sheet, trial example of pay sum from sheet
- see operations
- some refactoring

+++ src/by/epam/travel_agency/model/entity/User.java
@@ -22,7 +22,7 @@ public class User extends Entity {
     public User() {
     }
 
-    public User(long id, String login, String email, UserType role, boolean status) {
+    public User(int id, String login, String email, UserType role, boolean status) {
         super(id);
         this.login = login;
         this.email = email;
